use unified open in app approach, strip extra spaces in given url
user gets list of app to open a file. app has certain number. User has to chose and hit enter to open file or just hit enter to use first one from the list. Application list  is in the config, so it can be extended. System is checking if app is callable with the "which" command. Non-callable are going to be taken out from the list